refactor(provider): merge list-available into list --available
