--- cmd/provider/list_available.go
@@ -1,94 +1,0 @@
-package provider
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"strings"
-
-	"github.com/devsy-org/devsy/cmd/flags"
-	"github.com/devsy-org/devsy/pkg/config"
-	devsyhttp "github.com/devsy-org/devsy/pkg/http"
-	"github.com/devsy-org/devsy/pkg/table"
-	"github.com/spf13/cobra"
-)
-
-// ListAvailableCmd holds the list cmd flags.
-type ListAvailableCmd struct {
-	*flags.GlobalFlags
-}
-
-// NewListAvailableCmd creates a new command.
-func NewListAvailableCmd(flags *flags.GlobalFlags) *cobra.Command {
-	cmd := &ListAvailableCmd{
-		GlobalFlags: flags,
-	}
-	listAvailableCmd := &cobra.Command{
-		Use:   "list-available",
-		Short: "List providers available for installation",
-		Args:  cobra.NoArgs,
-		RunE: func(cobraCmd *cobra.Command, args []string) error {
-			return cmd.Run(cobraCmd.Context())
-		},
-	}
-
-	return listAvailableCmd
-}
-
-// Run runs the command logic.
-func (cmd *ListAvailableCmd) Run(ctx context.Context) error {
-	jsonResult, err := fetchProviderRepos(ctx)
-	if err != nil {
-		return err
-	}
-
-	_, _ = fmt.Fprintln(os.Stdout, "List of available providers from "+config.RepoOwner+":")
-	var rows [][]string
-	for _, v := range jsonResult {
-		name, ok := v["name"].(string)
-		if !ok || name == "" {
-			continue
-		}
-		if after, ok0 := strings.CutPrefix(name, config.ProviderPrefix); ok0 {
-			rows = append(rows, []string{after})
-		}
-	}
-	table.Print([]string{"Provider"}, rows)
-
-	return nil
-}
-
-func fetchProviderRepos(ctx context.Context) ([]map[string]any, error) {
-	req, err := http.NewRequestWithContext(ctx,
-		"GET",
-		config.GitHubAPIUserURL+"/repos?per_page=100",
-		nil,
-	)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := devsyhttp.GetHTTPClient().Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	result, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, fmt.Errorf("unexpected status code %d: %s", resp.StatusCode, string(result))
-	}
-
-	var jsonResult []map[string]any
-	if err := json.Unmarshal(result, &jsonResult); err != nil {
-		return nil, err
-	}
-
-	return jsonResult, nil
-}
