+++ cmd/provider/provider.go
@@ -17,7 +17,6 @@ func NewProviderCmd(flags *flags.GlobalFlags) *cobra.Command {
 	providerCmd.AddCommand(NewDefaultCmd(flags))
 	providerCmd.AddCommand(NewRemoveCmd(flags))
 	providerCmd.AddCommand(NewListCmd(flags))
-	providerCmd.AddCommand(NewListAvailableCmd(flags))
 	providerCmd.AddCommand(NewOptionsCmd(flags))
 	providerCmd.AddCommand(NewRenameCmd(flags))
 	providerCmd.AddCommand(NewSetOptionsCmd(flags))
